Consolidate suites, run async when possible
