+++ test-clients/src/main/java/com/hedera/services/bdd/suites/HapiApiSuite.java
@@ -9,9 +9,9 @@ package com.hedera.services.bdd.suites;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,12 +20,10 @@ package com.hedera.services.bdd.suites;
  * ‍
  */
 
-import com.google.common.math.Stats;
 import com.hedera.services.bdd.spec.HapiApiSpec;
 import com.hedera.services.bdd.spec.HapiSpecOperation;
 import com.hedera.services.bdd.spec.HapiSpecSetup;
 import com.hedera.services.bdd.spec.infrastructure.HapiApiClients;
-import com.hedera.services.bdd.spec.queries.HapiQueryOp;
 import com.hederahashgraph.api.proto.java.Key;
 import com.hederahashgraph.api.proto.java.KeyList;
 import org.apache.logging.log4j.Logger;
@@ -33,7 +31,6 @@ import org.apache.logging.log4j.Logger;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 import java.util.Random;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
@@ -48,6 +45,7 @@ import static com.hedera.services.bdd.suites.HapiApiSuite.FinalOutcome.SUITE_PAS
 
 public abstract class HapiApiSuite {
 	protected abstract Logger getResultsLogger();
+
 	protected abstract List<HapiApiSpec> getSpecsInSuite();
 
 	private static final Random r = new Random();
@@ -68,7 +66,10 @@ public abstract class HapiApiSuite {
 	public List<HapiApiSpec> getFinalSpecs() {
 		return finalSpecs;
 	}
-	public boolean leaksState() { return false; }
+
+	public boolean canRunAsync() {
+		return false;
+	}
 
 	public static final Key EMPTY_KEY = Key.newBuilder().setKeyList(KeyList.newBuilder().build()).build();
 	public static final String NONSENSE_KEY = "Jabberwocky!";
@@ -109,7 +110,8 @@ public abstract class HapiApiSuite {
 				: simpleName.substring(0, simpleName.length() - "Suite".length());
 	}
 
-	public enum FinalOutcome { SUITE_PASSED, SUITE_FAILED }
+	public enum FinalOutcome {SUITE_PASSED, SUITE_FAILED}
+
 	protected FinalOutcome finalOutcomeFor(List<HapiApiSpec> completedSpecs) {
 		return completedSpecs.stream().allMatch(HapiApiSpec::OK) ? SUITE_PASSED : SUITE_FAILED;
 	}
@@ -117,6 +119,7 @@ public abstract class HapiApiSuite {
 	public FinalOutcome runSuiteAsync() {
 		return runSuite(this::runAsync);
 	}
+
 	public FinalOutcome runSuiteSync() {
 		return runSuite(this::runSync);
 	}
@@ -136,20 +139,19 @@ public abstract class HapiApiSuite {
 		return Stream
 				.of(ops)
 				.map(op -> (op instanceof HapiSpecOperation)
-						? new HapiSpecOperation[] { (HapiSpecOperation)op }
-						: ((op instanceof List) ? ((List)op).toArray(new HapiSpecOperation[0]) : (HapiSpecOperation[])op))
+						? new HapiSpecOperation[] { (HapiSpecOperation) op }
+						: ((op instanceof List) ? ((List) op).toArray(
+						new HapiSpecOperation[0]) : (HapiSpecOperation[]) op))
 				.flatMap(Stream::of)
 				.toArray(n -> new HapiSpecOperation[n]);
 	}
+
 	protected List<HapiApiSpec> allOf(List<HapiApiSpec>... specLists) {
 		return Arrays.stream(specLists).flatMap(List::stream).collect(Collectors.toList());
 	}
 
 	protected HapiSpecOperation[] asOpArray(int N, Function<Integer, HapiSpecOperation> factory) {
 		return IntStream.range(0, N).mapToObj(i -> factory.apply(i)).toArray(n -> new HapiSpecOperation[n]);
-	}
-	protected HapiQueryOp<?>[] asQueryOpArray(int N, Function<Integer, HapiQueryOp<?>> factory) {
-		return IntStream.range(0, N).mapToObj(i -> factory.apply(i)).toArray(n -> new HapiQueryOp<?>[n]);
 	}
 
 	private void summarizeResults(Logger log) {
@@ -159,22 +161,9 @@ public abstract class HapiApiSuite {
 		}
 	}
 
-	private String asNormalApproximation(Stats stats) {
-		if (stats.count() > 1) {
-			return "~ \uD835\uDCA9(μ=" + rounded(stats.mean())
-					+ "ms, σ=" + rounded(stats.sampleStandardDeviation()) + "ms)";
-		} else {
-			return "= " + (long)stats.sum() + "ms";
-		}
-	}
-
-	private String rounded(double v) {
-		return String.format("%.2f", v);
-	}
-
 	private void runSync(Iterable<HapiApiSpec> specs) {
 		StreamSupport
-			.stream(specs.spliterator(), false)
+				.stream(specs.spliterator(), false)
 				.forEach(Runnable::run);
 	}
 
@@ -184,9 +173,5 @@ public abstract class HapiApiSuite {
 				.map(CompletableFuture::runAsync)
 				.toArray(n -> new CompletableFuture[n]);
 		CompletableFuture.allOf(futures).join();
-	}
-
-	public static boolean cacheRecordsAreAddedToState() {
-		return "true".equals(Optional.ofNullable(System.getenv("ADD_CACHE_RECORD_TO_STATE")).orElse(""));
 	}
 }
