+++ test-clients/src/main/java/com/hedera/services/bdd/suites/contract/ContractGetBytecodeSuite.java
@@ -61,6 +61,11 @@ public class ContractGetBytecodeSuite extends HapiApiSuite {
 		);
 	}
 
+	@Override
+	public boolean canRunAsync() {
+		return true;
+	}
+
 	private List<HapiApiSpec> negativeSpecs() {
 		return Arrays.asList(
 			invalidContractFromCostAnswer(),
