+++ test-clients/src/main/java/com/hedera/services/bdd/suites/contract/ContractUpdateSuite.java
@@ -49,7 +49,12 @@ public class ContractUpdateSuite extends HapiApiSuite {
 			Long.parseLong(HapiSpecSetup.getDefaultNodeProps().get("entities.maxLifetime"));
 
 	public static void main(String... args) {
-		new ContractUpdateSuite().runSuiteSync();
+		new ContractUpdateSuite().runSuiteAsync();
+	}
+
+	@Override
+	public boolean canRunAsync() {
+		return true;
 	}
 
 	@Override
