+++ test-clients/src/main/java/com/hedera/services/bdd/suites/contract/SmartContractFailFirstSpec.java
@@ -21,27 +21,11 @@ package com.hedera.services.bdd.suites.contract;
  */
 
 import com.hedera.services.bdd.spec.HapiApiSpec;
-import com.hedera.services.bdd.spec.infrastructure.meta.ContractResources;
-import com.hedera.services.bdd.spec.utilops.CustomSpecAssert;
 import com.hedera.services.bdd.suites.HapiApiSuite;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.List;
-
-import static com.hedera.services.bdd.spec.HapiApiSpec.defaultHapiSpec;
-import static com.hedera.services.bdd.spec.assertions.AccountInfoAsserts.changeFromSnapshot;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountBalance;
-import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTxnRecord;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCall;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.fileCreate;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.balanceSnapshot;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.CONTRACT_REVERT_EXECUTED;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INSUFFICIENT_GAS;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SUCCESS;
 
 public class SmartContractFailFirstSpec extends HapiApiSuite  {
 	private static final Logger log = LogManager.getLogger(SmartContractFailFirstSpec.class);
@@ -55,124 +39,9 @@ public class SmartContractFailFirstSpec extends HapiApiSuite  {
 	@Override
 	protected List<HapiApiSpec> getSpecsInSuite() {
 		return List.of(new HapiApiSpec[] {
-				smartContractFailFirst(),
 		});
 	}
 
-	HapiApiSpec smartContractFailFirst() {
-
-		return defaultHapiSpec("smartContractFailFirst")
-				.given(
-						cryptoCreate("payer").balance(1_000_000_000_000L).logged(),
-						fileCreate("bytecode")
-								.path(ContractResources.SIMPLE_STORAGE_BYTECODE_PATH)
-				).when(
-						withOpContext((spec, ignore) -> {
-							var subop1 = balanceSnapshot("balanceBefore0", "payer");
-
-							var subop2 =
-									contractCreate("failInsufficientGas")
-											.balance(0)
-											.payingWith("payer")
-											.gas(1)
-											.bytecode("bytecode")
-											.hasKnownStatus(INSUFFICIENT_GAS)
-											.via("failInsufficientGas");
-
-							var subop3 = getTxnRecord("failInsufficientGas");
-							CustomSpecAssert.allRunFor(spec, subop1, subop2, subop3);
-							long delta = subop3.getResponseRecord().getTransactionFee();
-
-							var subop4 = getAccountBalance("payer").hasTinyBars(changeFromSnapshot("balanceBefore0", -delta));
-							CustomSpecAssert.allRunFor(spec,  subop4);
-
-						}),
-
-
-
-						withOpContext((spec, ignore) -> {
-							var subop1 = balanceSnapshot("balanceBefore1", "payer");
-
-							var subop2 = contractCreate("failInvalidInitialBalance")
-									.balance(100_000_000_000L)
-									.payingWith("payer")
-									.gas(250_000L)
-									.bytecode("bytecode")
-									.via("failInvalidInitialBalance")
-									.hasKnownStatus(CONTRACT_REVERT_EXECUTED);
-
-							var subop3 = getTxnRecord("failInvalidInitialBalance");
-							CustomSpecAssert.allRunFor(spec, subop1, subop2, subop3);
-							long delta = subop3.getResponseRecord().getTransactionFee();
-
-							var subop4 = getAccountBalance("payer").hasTinyBars(changeFromSnapshot("balanceBefore1", -delta));
-							CustomSpecAssert.allRunFor(spec,  subop4);
-
-						}),
-
-
-						withOpContext((spec, ignore) -> {
-							var subop1 = balanceSnapshot("balanceBefore2", "payer");
-
-							var subop2 = contractCreate("successWithZeroInitialBalance")
-									.balance(0L)
-									.payingWith("payer")
-									.gas(250_000L)
-									.bytecode("bytecode")
-									.hasKnownStatus(SUCCESS)
-									.via("successWithZeroInitialBalance");
-
-							var subop3 = getTxnRecord("successWithZeroInitialBalance");
-							CustomSpecAssert.allRunFor(spec, subop1, subop2, subop3);
-							long delta = subop3.getResponseRecord().getTransactionFee();
-
-							var subop4 = getAccountBalance("payer").hasTinyBars(changeFromSnapshot("balanceBefore2", -delta));
-							CustomSpecAssert.allRunFor(spec, subop4);
-
-						}),
-
-						withOpContext((spec, ignore) -> {
-							var subop1 = balanceSnapshot("balanceBefore3", "payer");
-
-							var subop2 = contractCall("successWithZeroInitialBalance", ContractResources.SIMPLE_STORAGE_SETTER_ABI, 999_999L )
-									.payingWith("payer")
-									.gas(300_000L)
-									.hasKnownStatus(SUCCESS)
-									.via("setValue");
-
-							var subop3 = getTxnRecord("setValue");
-							CustomSpecAssert.allRunFor(spec, subop1, subop2, subop3);
-							long delta = subop3.getResponseRecord().getTransactionFee();
-
-							var subop4 = getAccountBalance("payer").hasTinyBars(changeFromSnapshot("balanceBefore3", -delta));
-							CustomSpecAssert.allRunFor(spec, subop4);
-
-						}),
-
-
-						withOpContext((spec, ignore) -> {
-							var subop1 = balanceSnapshot("balanceBefore4", "payer");
-
-							var subop2 = contractCall("successWithZeroInitialBalance", ContractResources.SIMPLE_STORAGE_GETTER_ABI)
-									.payingWith("payer")
-									.gas(300_000L)
-									.hasKnownStatus(SUCCESS)
-									.via("getValue");
-
-							var subop3 = getTxnRecord("getValue");
-							CustomSpecAssert.allRunFor(spec, subop1, subop2, subop3);
-							long delta = subop3.getResponseRecord().getTransactionFee();
-
-							var subop4 = getAccountBalance("payer").hasTinyBars(changeFromSnapshot("balanceBefore4", -delta));
-							CustomSpecAssert.allRunFor(spec, subop4);
-
-						})
-						).then(
-						getTxnRecord("failInsufficientGas"),
-						getTxnRecord("successWithZeroInitialBalance"),
-						getTxnRecord("failInvalidInitialBalance")
-				);
-	}
 
 	@Override
 	protected Logger getResultsLogger() {
