+++ test-clients/src/main/java/com/hedera/services/bdd/suites/fees/TransferListServiceFeesSuite.java
@@ -57,7 +57,7 @@ public class TransferListServiceFeesSuite extends HapiApiSuite {
 	}
 
 	@Override
-	public boolean leaksState() {
+	public boolean canRunAsync() {
 		return true;
 	}
