+++ test-clients/src/main/java/com/hedera/services/bdd/suites/file/FileCreateSuite.java
@@ -55,7 +55,12 @@ public class FileCreateSuite extends HapiApiSuite {
 			Long.parseLong(HapiSpecSetup.getDefaultNodeProps().get("entities.maxLifetime"));
 
 	public static void main(String... args) {
-		new FileCreateSuite().runSuiteSync();
+		new FileCreateSuite().runSuiteAsync();
+	}
+
+	@Override
+	public boolean canRunAsync() {
+		return true;
 	}
 
 	@Override
