+++ test-clients/src/main/java/com/hedera/services/bdd/suites/issues/Issue1742Suite.java
@@ -50,7 +50,7 @@ public class Issue1742Suite extends HapiApiSuite {
 	}
 
 	@Override
-	public boolean leaksState() {
+	public boolean canRunAsync() {
 		return true;
 	}
