+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/contract/ContractCallLocalPerfSuite.java
@@ -52,7 +52,7 @@ public class ContractCallLocalPerfSuite extends HapiApiSuite {
 	}
 
 	@Override
-	public boolean leaksState() {
+	public boolean canRunAsync() {
 		return true;
 	}
