+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/contract/ContractCallPerfSuite.java
@@ -57,7 +57,7 @@ public class ContractCallPerfSuite extends HapiApiSuite {
 	}
 
 	@Override
-	public boolean leaksState() {
+	public boolean canRunAsync() {
 		return true;
 	}
