+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/crypto/CryptoTransferPerfSuite.java
@@ -52,7 +52,7 @@ public class CryptoTransferPerfSuite extends HapiApiSuite {
 	}
 
 	@Override
-	public boolean leaksState() {
+	public boolean canRunAsync() {
 		return true;
 	}
