+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/topic/CreateTopicPerfSuite.java
@@ -53,7 +53,7 @@ public class CreateTopicPerfSuite extends HapiApiSuite {
 	}
 
 	@Override
-	public boolean leaksState() {
+	public boolean canRunAsync() {
 		return true;
 	}
