+++ test-clients/src/main/java/com/hedera/services/bdd/suites/perf/topic/SubmitMessagePerfSuite.java
@@ -51,7 +51,7 @@ public class SubmitMessagePerfSuite extends HapiApiSuite {
 	}
 
 	@Override
-	public boolean leaksState() {
+	public boolean canRunAsync() {
 		return true;
 	}
