+++ test-clients/src/main/java/com/hedera/services/bdd/suites/token/TokenAssociationSpecs.java
@@ -86,7 +86,7 @@ public class TokenAssociationSpecs extends HapiApiSuite {
 	private static final String VANILLA_TOKEN = "TokenD";
 
 	public static void main(String... args) {
-		new TokenAssociationSpecs().runSuiteSync();
+		new TokenAssociationSpecs().runSuiteAsync();
 	}
 
 	@Override
@@ -104,6 +104,11 @@ public class TokenAssociationSpecs extends HapiApiSuite {
 						contractInfoQueriesAsExpected(),
 				}
 		);
+	}
+
+	@Override
+	public boolean canRunAsync() {
+		return true;
 	}
 
 	public HapiApiSpec handlesUseOfDefaultTokenId() {
